1. implemented peers list load. added peer sort by ping. iteration 3

+++ app/src/main/java/io/github/chronosx88/yggdrasil/models/PeerInfo.kt
@@ -17,7 +17,7 @@ class PeerInfo {
     var address: InetAddress
     var port = 0
     var countryCode: String
-    var ping: Float = Float.MAX_VALUE
+    var ping: Int = Int.MAX_VALUE
 
     override fun toString(): String {
         var a = address.toString();
